Move Season & Shows to Performance Subdomain
