+++ packages/web/src/domains/performance/show/__template__/ShowTemplate.tsx
