+++ packages/web/src/domains/performance/show/__template__/components/ActionBar/ActionBar.tsx
