+++ packages/web/src/domains/performance/show/__template__/components/Hero/Hero.styles.ts
