+++ packages/web/src/domains/performance/show/components/ShowPoster/ShowPoster.tsx
