+++ packages/web/src/domains/performance/show/components/ShowPoster/__styles.ts
