+++ packages/web/src/domains/performance/show/components/ShowPosterGrid/__styles.ts
